requirements.txt
scapy
colorama
python-nmap
netifaces
manuf
